% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCCHeatmap.R
\name{CCCHeatmap}
\alias{CCCHeatmap}
\title{Cell-cell Communiction Heatmap}
\usage{
CCCHeatmap(
  restab,
  ligand = NULL,
  receptor = NULL,
  pathway = NULL,
  celltype = NULL,
  bound = 1,
  cutoff = 0.2,
  ...
)
}
\arguments{
\item{restab}{A \code{data.frame} returned by \code{FRETCOM()}.}

\item{ligand}{A specific ligand name in \code{restab}, should be paired with \code{receptor}.}

\item{receptor}{A specific receptor name in \code{restab}, should be paired with \code{ligand}.}

\item{pathway}{A specific pathway name in \code{restab} (if any). Once \code{pathway} is
defined, \code{ligand} and \code{receptor} will not be considered.}

\item{celltype}{A vector including all cell type names that the user wants to
include in the heatmap. If \code{celltype=NULL}, the cell types in
\code{restab} will be used. The default is \code{NULL}.}

\item{bound}{A numeric value indicates the lower and upper bounds of the
coefficient estimates displayed in the heatmap. All values
greater than \code{abs(bound)} will be truncated to \code{abs(bound)},
and all values smaller than \code{-abs(bound)} will be truncated
to \code{-abs(bound)}. The default is \code{1}.}

\item{cutoff}{A numeric value within the range of \verb{(0, 1)} indicates the cutoff
for the adjusted p-values. If the adjusted p-value is less than
the specified cutoff, we treat it as significant, and it will be
displayed in the heatmap. The default is \code{0.2}.}

\item{...}{Other arguments passed on to \code{ComplexHeatmap::draw()}.}
}
\value{
A heatmap produced by \code{ComplexHeatmap}.
}
\description{
Cell-cell Communiction Heatmap
}
