% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypeMatrix.R
\name{CellTypeMatrix}
\alias{CellTypeMatrix}
\title{Cell Type Identification Matrix}
\usage{
CellTypeMatrix(stmeta, celltype)
}
\arguments{
\item{stmeta}{A \code{data.frame} contains cell type information for each cell, where
each row represents a cell. Please make sure that the order of rows
corresponds to the columns of the gene expression matrix.}

\item{celltype}{A variable name that represents the cell type information in \code{stmata}.}
}
\value{
A cell type identification matrix.
}
\description{
This function transforms the cell type information for each cell into an
identification matrix that can be used in our downstream function \code{FRETCOM()}.
}
