% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRETCOM.R
\name{FRETCOM}
\alias{FRETCOM}
\title{Frequentist Tweedie Modeling of Cell-cell Communication}
\usage{
FRETCOM(
  stdat,
  M,
  ligands,
  receptors,
  ngridx = 5,
  ngridy = 5,
  coordx,
  coordy,
  maxdist,
  sampleid = NULL,
  rholist = c(0.2, 0.5, 0.8),
  parallel = FALSE,
  ncores = 4
)
}
\arguments{
\item{stdat}{A normalized gene expression matrix where each row represents a
gene and each column represents a cell/spot. Please make sure row
names are gene names.}

\item{M}{A cell type proportion/identification matrix where each row represents
a cell/spot and each column represents a cell type. Please ensure that
the row names (and order) match the column names (and order) of \code{stdat}.}

\item{ligands}{A list or vector of gene names for ligands corresponding to LR
pairs. Please make sure the order is the same as \code{receptors}.}

\item{receptors}{A list or vector of gene names for receptors corresponding to LR
pairs. Please make sure the order is the same as \code{ligands}.}

\item{ngridx}{An integer to indicate the number of grids that will be create
for the X coordinate if \code{re=TRUE}. Default is \code{5}.}

\item{ngridy}{An integer to indicate the number of grids that will be create
for the Y coordinate if \code{re=TRUE}. Default is \code{5}.}

\item{coordx}{A vector of X coordinates for cells/spots. Please make sure the
order can match the order of columns in \code{stdat}.}

\item{coordy}{A vector of Y coordinates for cells/spots. Please make sure the
order can match the order of columns in \code{stdat}.}

\item{maxdist}{The maximum distance allowed for communication. This number
should be based on the units of \code{coordx} and \code{coordy}. Usually,
we would recommend 200-300 μm.}

\item{sampleid}{A vector of sample IDs for cells/spots. Please make sure the
order can match the order of columns in \code{stdat}. Default is \code{NULL},
i.e., only considering one sample.}

\item{rholist}{A vector that includes all possible rho values (communication
constraint parameter) that need to be tried in model fitting.
The best one will be automatically chosen based on AIC. Default
value is \code{c(0.2, 0.5, 0.8)}.}

\item{parallel}{A logical value to indicate whether or not use parallel computing.
Default is \code{FALSE}.}

\item{ncores}{An integer to indicate the number of cores to be used for parallel
computing. This is only considered if \code{parallel=TRUE}. Default is \code{4}.}
}
\value{
A \code{data.frame} contains the inference results of cell-cell communication.
}
\description{
Frequentist Tweedie Modeling of Cell-cell Communication
}
\references{
{
\insertRef{wu2023inferring}{TWCOM}
}
}
