% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MExample.R
\docType{data}
\name{MExample}
\alias{MExample}
\title{M Matrix}
\format{
An object of class \code{data.frame} with 1932 rows and 24 columns.
}
\usage{
MExample
}
\description{
The cell type proportion matrix of STdata using RCTD.
}
\references{
{
\insertRef{ji2020multimodal}{TWCOM}
\insertRef{cable2022robust}{TWCOM}
}
}
\keyword{datasets}
