% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STdata.R
\docType{data}
\name{STdata}
\alias{STdata}
\title{STdata}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10703 rows and 1932 columns.
}
\usage{
STdata
}
\description{
The gene expression matrix of SRT data.
}
\references{
{
\insertRef{ji2020multimodal}{TWCOM}
}
}
\keyword{datasets}
