% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STmeta.R
\docType{data}
\name{STmeta}
\alias{STmeta}
\title{STmeta}
\format{
An object of class \code{data.frame} with 1932 rows and 3 columns.
}
\usage{
STmeta
}
\description{
The X and Y coordinates information of SRT data.
}
\references{
{
\insertRef{ji2020multimodal}{TWCOM}
}
}
\keyword{datasets}
